Separate all commands into respective class files
