--- src/main/java/com/arcaneminecraft/bungee/command/PermissionCommands.java
@@ -1,9 +1,0 @@
-package com.arcaneminecraft.bungee.command;
-
-
-/**
- * Commands in this class:
- * badge, badgeadmin, greylist
- */
-public class PermissionCommands {
-}
